refactor(ml-1268): merge question controller-post into controller
Nothing else in the frontend codebase splits a feature into
controller.js + controller-post.js — the rest handle GET and POST in
one controller.js. Fold questionPostController back in alongside
questionController and merge the test files the same way (matching
the outcome dir convention). Lift the toArray helper into
question/utils.js with its own unit tests since it's a pure
branching helper.
